+++ tests/data/cases/combine-fstring.py
@@ -1,0 +1,7 @@
+sep= " "
+FSTRING = f"many" f"{sep}" f"tiny" f"{sep}" f"fstrings"
+
+# output
+
+sep = " "
+FSTRING = f"many{sep}tiny{sep}fstrings"